pythonn/requirements.txt
acme==1.21.0
blinker==1.8.2
cachetools==5.4.0
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Flask==3.0.3
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gunicorn==23.0.0
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
josepy==1.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
parsedatetime==2.6
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.0.3
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
